AL3Oct25_1
working on manus. found an odd element but looks right through all the slices. tried to visualise with volume rendering using different presets but none are great at visualising. wanted to return to default colume rendering but i could not find it. pushing and reopening through morphodepot to return that to default.