AL16Oct25_1
the number of segments were becoming unwieldy, so i separated all muscles segmented that were not associated with the forelimb to another segmentation node which i saved to my instance as "non-forelimb.seg.nrrd." the performance has been improved and i will finalize the muscles of strictly the forelimb in this segmentation issue. i can load in the work on the other muscles for a different issue focused on the non-forelimb musculature.